Issue #154 refactor: HistoryLocalDataSource
