+++ app/src/main/java/com/lateinit/rightweight/di/RepositoryModule.kt
@@ -46,9 +46,9 @@ class RepositoryModule {
     @Provides
     @Singleton
     fun getHistoryRepository(
-        historyLocalDataSource: HistoryLocalDataSource
+        historyLocalDataSourceImpl: HistoryLocalLocalDataSourceImpl
     ): HistoryRepository {
-        return HistoryRepositoryImpl(historyLocalDataSource)
+        return HistoryRepositoryImpl(historyLocalDataSourceImpl)
     }
 
     @Provides
